+++ openpyxl/reader/tests/test_worksheet.py
@@ -15,6 +15,7 @@ from openpyxl.xml.constants import SHEET_MAIN_NS
 from openpyxl.cell import Cell
 from openpyxl.utils.indexed_list import IndexedList
 from openpyxl.styles import Style
+from openpyxl.worksheet import Worksheet
 
 
 def test_get_xml_iter():
@@ -56,33 +57,9 @@ def test_get_xml_iter():
 
 
 @pytest.fixture
-def Worksheet(Workbook):
-    from openpyxl.styles import numbers
+def Workbook():
     from openpyxl.styles.style import StyleId
-    from openpyxl.worksheet.header_footer import HeaderFooter
-
-    class DummyWorkbook:
-
-        _guess_types = False
-        data_only = False
-
-        def __init__(self):
-            self.shared_strings = IndexedList()
-            self.shared_strings.add("hello world")
-            self.shared_styles = 28*[DummyStyle()]
-            self.shared_styles.append(Style())
-            self._fonts = IndexedList()
-            self._fills = IndexedList()
-            self._number_formats = IndexedList()
-            self._borders = IndexedList()
-            self._alignments = IndexedList()
-            self._protections = IndexedList()
-            self._cell_styles = IndexedList()
-            self.vba_archive = None
-            for i in range(29):
-                self._cell_styles.add((StyleId(i, i, i, i, i, i)))
-            self._cell_styles.add(StyleId(fillId=4, borderId=6, alignmentId=1, protectionId=0))
-
+    from openpyxl.styles import numbers
 
     class DummyStyle:
         number_format = numbers.FORMAT_GENERAL
@@ -93,61 +70,60 @@ def Worksheet(Workbook):
         protection = ""
 
         def copy(self, **kw):
-            return selfexc
+            return self
 
 
-    class DummyWorksheet:
+    class DummyWorkbook:
 
-        encoding = "utf-8"
-        title = "Dummy"
+        _guess_types = False
+        data_only = False
+        _colors = []
 
         def __init__(self):
-            self.parent = DummyWorkbook()
-            self.column_dimensions = {}
-            self.row_dimensions = {}
-            self._styles = {}
-            self.cell = None
-            self._cells = {}
-            self._data_validations = []
-            self.header_footer = HeaderFooter()
-            self.vba_controls = None
+            self._differential_styles = []
+            self.shared_strings = IndexedList()
+            self.shared_strings.add("hello world")
+            self._fonts = IndexedList()
+            self._fills = IndexedList()
+            self._number_formats = IndexedList()
+            self._borders = IndexedList()
+            self._alignments = IndexedList()
+            self._protections = IndexedList()
+            self._cell_styles = IndexedList()
+            self.vba_archive = None
+            for i in range(29):
+                self._cell_styles.add((StyleId(i, i, i, i, i, i)))
+            self._cell_styles.add(StyleId(fillId=4, borderId=6, alignmentId=1))
 
-        def _add_cell(self, cell):
-            self._cells[cell.coordinate] = cell
 
-        def __getitem__(self, value):
-            cell = self._cells.get(value)
+        def create_sheet(self, title):
+            return Worksheet(self)
 
-            if cell is None:
-                cell = Cell(self, 'A', 1)
-                self._cells[value] = cell
-            return cell
+        def get_sheet_names(self):
+            return []
 
-        def get_style(self, coordinate):
-            return DummyStyle()
-
-    return DummyWorksheet()
+    return DummyWorkbook()
 
 
 @pytest.fixture
-def WorkSheetParser(Worksheet):
+def WorkSheetParser(Workbook):
     """Setup a parser instance with an empty source"""
     from .. worksheet import WorkSheetParser
-    return WorkSheetParser(Worksheet, None, {0:'a'}, {})
+    return WorkSheetParser(Workbook, 'sheet', None, {0:'a'})
 
 
 @pytest.fixture
-def WorkSheetParserKeepVBA(Worksheet):
+def WorkSheetParserKeepVBA(Workbook):
     """Setup a parser instance with an empty source"""
-    Worksheet.parent.vba_archive=True
+    Workbook.vba_archive=True
     from .. worksheet import WorkSheetParser
-    return WorkSheetParser(Worksheet, None, {0:'a'}, {})
+    return WorkSheetParser(Workbook, "sheet", {0:'a'}, {})
 
 
-def test_col_width(datadir, Worksheet, WorkSheetParser):
+def test_col_width(datadir, WorkSheetParser):
     datadir.chdir()
-    ws = Worksheet
     parser = WorkSheetParser
+    ws = parser.ws
 
     with open("complex-styles-worksheet.xml", "rb") as src:
         cols = iterparse(src, tag='{%s}col' % SHEET_MAIN_NS)
@@ -160,10 +136,10 @@ def test_col_width(datadir, Worksheet, WorkSheetParser):
                                                'width': '31.1640625'}
 
 
-def test_hidden_col(datadir, Worksheet, WorkSheetParser):
+def test_hidden_col(datadir, WorkSheetParser):
     datadir.chdir()
-    ws = Worksheet
     parser = WorkSheetParser
+    ws = parser.ws
 
     with open("hidden_rows_cols.xml", "rb") as src:
         cols = iterparse(src, tag='{%s}col' % SHEET_MAIN_NS)
@@ -174,10 +150,11 @@ def test_hidden_col(datadir, Worksheet, WorkSheetParser):
                                                '1', 'max': '4', 'min': '4'}
 
 
-def test_styled_col(datadir, Worksheet, WorkSheetParser):
+def test_styled_col(datadir, WorkSheetParser):
     datadir.chdir()
-    ws = Worksheet
     parser = WorkSheetParser
+    ws = parser.ws
+
     with open("complex-styles-worksheet.xml", "rb") as src:
         cols = iterparse(src, tag='{%s}col' % SHEET_MAIN_NS)
         for _, col in cols:
@@ -188,10 +165,10 @@ def test_styled_col(datadir, Worksheet, WorkSheetParser):
     assert dict(cd) ==  {'customWidth': '1', 'max': '9', 'min': '9', 'width': '25', 'style':'28'}
 
 
-def test_hidden_row(datadir, Worksheet, WorkSheetParser):
+def test_hidden_row(datadir, WorkSheetParser):
     datadir.chdir()
-    ws = Worksheet
     parser = WorkSheetParser
+    ws = parser.ws
 
     with open("hidden_rows_cols.xml", "rb") as src:
         rows = iterparse(src, tag='{%s}row' % SHEET_MAIN_NS)
@@ -201,12 +178,11 @@ def test_hidden_row(datadir, Worksheet, WorkSheetParser):
     assert dict(ws.row_dimensions[2]) == {'hidden': '1'}
 
 
-def test_styled_row(datadir, Worksheet, WorkSheetParser):
+def test_styled_row(datadir, WorkSheetParser):
     datadir.chdir()
-    ws = Worksheet
     parser = WorkSheetParser
+    ws = parser.ws
     parser.shared_strings = dict((i, i) for i in range(30))
-    parser.style_table = ws.parent.shared_styles
 
     with open("complex-styles-worksheet.xml", "rb") as src:
         rows = iterparse(src, tag='{%s}row' % SHEET_MAIN_NS)
@@ -219,10 +195,10 @@ def test_styled_row(datadir, Worksheet, WorkSheetParser):
     assert dict(rd) == {'s':'28', 'customFormat':'1'}
 
 
-def test_sheet_protection(datadir, Worksheet, WorkSheetParser):
+def test_sheet_protection(datadir, WorkSheetParser):
     datadir.chdir()
-    ws = Worksheet
     parser = WorkSheetParser
+    ws = parser.ws
 
     with open("protected_sheet.xml", "rb") as src:
         tree = iterparse(src, tag='{%s}sheetProtection' % SHEET_MAIN_NS)
@@ -238,9 +214,9 @@ def test_sheet_protection(datadir, Worksheet, WorkSheetParser):
     }
 
 
-def test_formula_without_value(Worksheet, WorkSheetParser):
-    ws = Worksheet
+def test_formula_without_value(WorkSheetParser):
     parser = WorkSheetParser
+    ws = parser.ws
 
     src = """
       <x:c r="A1" xmlns:x="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
@@ -255,9 +231,9 @@ def test_formula_without_value(Worksheet, WorkSheetParser):
     assert ws['A1'].value == '=IF(TRUE, "y", "n")'
 
 
-def test_formula(Worksheet, WorkSheetParser):
-    ws = Worksheet
+def test_formula(WorkSheetParser):
     parser = WorkSheetParser
+    ws = parser.ws
 
     src = """
     <x:c r="A1" t="str" xmlns:x="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
@@ -272,9 +248,9 @@ def test_formula(Worksheet, WorkSheetParser):
     assert ws['A1'].value == '=IF(TRUE, "y", "n")'
 
 
-def test_formula_data_only(Worksheet, WorkSheetParser):
-    ws = Worksheet
+def test_formula_data_only(WorkSheetParser):
     parser = WorkSheetParser
+    ws = parser.ws
     parser.data_only = True
 
     src = """
@@ -290,9 +266,9 @@ def test_formula_data_only(Worksheet, WorkSheetParser):
     assert ws['A1'].value == 3
 
 
-def test_string_formula_data_only(Worksheet, WorkSheetParser):
-    ws = Worksheet
+def test_string_formula_data_only(WorkSheetParser):
     parser = WorkSheetParser
+    ws = parser.ws
     parser.data_only = True
 
     src = """
@@ -308,9 +284,9 @@ def test_string_formula_data_only(Worksheet, WorkSheetParser):
     assert ws['A1'].value == 'y'
 
 
-def test_number(Worksheet, WorkSheetParser):
-    ws = Worksheet
+def test_number(WorkSheetParser):
     parser = WorkSheetParser
+    ws = parser.ws
 
     src = """
     <x:c r="A1" xmlns:x="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
@@ -324,9 +300,9 @@ def test_number(Worksheet, WorkSheetParser):
     assert ws['A1'].value == 1
 
 
-def test_string(Worksheet, WorkSheetParser):
-    ws = Worksheet
+def test_string(WorkSheetParser):
     parser = WorkSheetParser
+    ws = parser.ws
 
     src = """
     <x:c r="A1" t="s" xmlns:x="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
@@ -340,9 +316,9 @@ def test_string(Worksheet, WorkSheetParser):
     assert ws['A1'].value == "a"
 
 
-def test_boolean(Worksheet, WorkSheetParser):
-    ws = Worksheet
+def test_boolean(WorkSheetParser):
     parser = WorkSheetParser
+    ws = parser.ws
 
     src = """
     <x:c r="A1" t="b" xmlns:x="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
@@ -356,10 +332,9 @@ def test_boolean(Worksheet, WorkSheetParser):
     assert ws['A1'].value is True
 
 
-def test_inline_string(Worksheet, WorkSheetParser, datadir):
-    ws = Worksheet
+def test_inline_string(WorkSheetParser, datadir):
     parser = WorkSheetParser
-    parser.style_table = ws.parent.shared_styles
+    ws = parser.ws
     datadir.chdir()
 
     with open("Table1-XmlFromAccess.xml") as src:
@@ -371,10 +346,9 @@ def test_inline_string(Worksheet, WorkSheetParser, datadir):
     assert ws['A1'].value == "ID"
 
 
-def test_inline_richtext(Worksheet, WorkSheetParser, datadir):
-    ws = Worksheet
+def test_inline_richtext(WorkSheetParser, datadir):
     parser = WorkSheetParser
-    parser.style_table = ws.parent.shared_styles
+    ws = parser.ws
     datadir.chdir()
     with open("jasper_sheet.xml", "rb") as src:
         sheet = fromstring(src.read())
@@ -387,9 +361,9 @@ def test_inline_richtext(Worksheet, WorkSheetParser, datadir):
     assert cell.value == "11 de September de 2014"
 
 
-def test_data_validation(Worksheet, WorkSheetParser, datadir):
-    ws = Worksheet
+def test_data_validation(WorkSheetParser, datadir):
     parser = WorkSheetParser
+    ws = parser.ws
     datadir.chdir()
 
     with open("worksheet_data_validation.xml") as src: