+++ openpyxl/worksheet/relationship.py
@@ -1,27 +1,29 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2015 openpyxl
 
+from openpyxl.descriptors import String, Set, NoneSet
+from openpyxl.descriptors.serialisable import Serialisable
 
 from openpyxl.xml.constants import REL_NS, PKG_REL_NS
 from openpyxl.xml.functions import Element, SubElement, tostring
 
-class Relationship(object):
+
+class Relationship(Serialisable):
     """Represents many kinds of relationships."""
     # TODO: Use this object for workbook relationships as well as
     # worksheet relationships
 
-    TYPES = ("hyperlink", "drawing", "image")
+    tagname = "Relationship"
 
-    def __init__(self, rel_type, target=None, target_mode=None, id=None):
-        if rel_type not in self.TYPES:
-            raise ValueError("Invalid relationship type %s" % rel_type)
-        self.type = "%s/%s" % (REL_NS, rel_type)
+    type = String()
+    target = String()
+    targetMode = String(allow_none=True)
+    id = String()
+
+
+    def __init__(self, type, target=None, targetMode=None, id=None):
+        self.type = "%s/%s" % (REL_NS, type)
         self.target = target
-        self.target_mode = target_mode
+        self.targetMode = targetMode
         self.id = id
-
-    def __repr__(self):
-        root = Element("{%s}Relationships" % PKG_REL_NS)
-        SubElement(root, "{%s}Relationship" % PKG_REL_NS, self.__dict__)
-        return tostring(root)