+++ openpyxl/writer/etree_worksheet.py
@@ -1,54 +1,45 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2015 openpyxl
 
-from openpyxl.compat import iterkeys, itervalues, safe_string
+from operator import itemgetter
 
-from openpyxl.utils import COORD_RE, column_index_from_string
+from openpyxl.compat import safe_string
 from openpyxl.xml.functions import xmlfile, Element, SubElement
-
-
-def row_sort(cell):
-    """Translate column names for sorting."""
-    return column_index_from_string(cell.column)
 
 
 def get_rows_to_write(worksheet):
     """Return all rows, and any cells that they contain"""
-    # Ensure a blank cell exists if it has a style
-    for styleCoord in iterkeys(worksheet._styles):
-        if isinstance(styleCoord, str) and COORD_RE.search(styleCoord):
-            worksheet.cell(styleCoord)
+    # order cells by row
+    rows = {}
+    for (row, col), cell in worksheet._cells.items():
+        rows.setdefault(row, []).append((col, cell))
 
-    # create rows of cells
-    cells_by_row = {}
-    for cell in itervalues(worksheet._cells):
-        cells_by_row.setdefault(cell.row, []).append(cell)
-
-    # make sure rows that only have a height set are returned
+    # add empty rows if styling has been applied
     for row_idx in worksheet.row_dimensions:
-        if row_idx not in cells_by_row:
-            cells_by_row[row_idx] = []
+        if row_idx not in rows:
+            rows[row_idx] = []
 
-    return cells_by_row
+    return sorted(rows.items())
 
 
 def write_rows(xf, worksheet):
     """Write worksheet data to xml."""
 
-    cells_by_row = get_rows_to_write(worksheet)
+    all_rows = get_rows_to_write(worksheet)
+
+    dims = worksheet.row_dimensions
+    max_column = worksheet.max_column
 
     with xf.element("sheetData"):
-        for row_idx in sorted(cells_by_row):
-            # row meta data
-            row_dimension = worksheet.row_dimensions[row_idx]
-            attrs = {'r': '%d' % row_idx,
-                     'spans': '1:%d' % worksheet.max_column}
-            attrs.update(dict(row_dimension))
+        for row_idx, row in all_rows:
+
+            attrs = {'r': '%d' % row_idx, 'spans': '1:%d' % max_column}
+            if row_idx in dims:
+                row_dimension = dims[row_idx]
+                attrs.update(dict(row_dimension))
 
             with xf.element("row", attrs):
-
-                row_cells = cells_by_row[row_idx]
-                for cell in sorted(row_cells, key=row_sort):
+                for col, cell in sorted(row, key=itemgetter(0)):
                     el = write_cell(worksheet, cell)
                     xf.write(el)