+++ openpyxl/writer/lxml_worksheet.py
@@ -1,61 +1,33 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2015 openpyxl
 
-from io import BytesIO
+from operator import itemgetter
 
-from openpyxl.compat import (
-    itervalues,
-    safe_string,
-    iteritems
-)
-from openpyxl.cell import (
-    column_index_from_string,
-    coordinate_from_string,
-)
-from openpyxl.xml.constants import (
-    REL_NS,
-    SHEET_MAIN_NS
-)
+from openpyxl.compat import safe_string
 
-from openpyxl.formatting import ConditionalFormatting
-from openpyxl.worksheet.datavalidation import writer
-
-from .worksheet import (
-    write_datavalidation,
-    write_properties,
-    write_format,
-    write_cols,
-    write_autofilter,
-    write_mergecells,
-    write_conditional_formatting,
-    write_header_footer,
-    write_hyperlinks,
-    write_pagebreaks,
-)
-
-from .etree_worksheet import get_rows_to_write, row_sort
-from openpyxl.xml.functions import xmlfile, Element, SubElement
-
+from .etree_worksheet import get_rows_to_write
+from openpyxl.xml.functions import xmlfile
 
 ### LXML optimisation using xf.element to reduce instance creation
 
 def write_rows(xf, worksheet):
     """Write worksheet data to xml."""
 
-    cells_by_row = get_rows_to_write(worksheet)
+    all_rows = get_rows_to_write(worksheet)
+
+    dims = worksheet.row_dimensions
+    max_column = worksheet.max_column
 
     with xf.element("sheetData"):
-        for row_idx in sorted(cells_by_row):
-            # row meta data
-            row_dimension = worksheet.row_dimensions[row_idx]
-            attrs = {'r': '%d' % row_idx,
-                     'spans': '1:%d' % worksheet.max_column}
-            attrs.update(dict(row_dimension))
+        for row_idx, row in sorted(all_rows):
 
+            attrs = {'r': '%d' % row_idx, 'spans': '1:%d' % max_column}
+            if row_idx in dims:
+                row_dimension = dims[row_idx]
+                attrs.update(dict(row_dimension))
             with xf.element("row", attrs):
 
-                row_cells = cells_by_row[row_idx]
-                for cell in sorted(row_cells, key=row_sort):
+                for col, cell in sorted(row, key=itemgetter(0)):
                     if cell.value is None and not cell.has_style:
                         continue
                     write_cell(xf, worksheet, cell)