+++ doc/source/changes.rst
@@ -1,3 +1,20 @@
+2.3.0 (unreleased)
+==================
+
+
+Major changes
+-------------
+
+* Shift to using (row, column) indexing for cells. Cells will at some point *lose* coordinates.
+* New implementation of conditional formatting. Databars now partially preserved.
+
+
+Minor changes
+-------------
+
+* Read-only and write-only worksheets renamed.
+
+
 2.2.1 (unreleased)
 ==================
 
@@ -21,7 +38,6 @@ Bug fixes
 * `#440 <https://bitbucket.org/openpyxl/openpyxl/issue/440>`_ Error when customised versions of builtin formats
 * `#442 <https://bitbucket.org/openpyxl/openpyxl/issue/442>`_ Exception raised when a fill element contains no children
 * `#444 <https://bitbucket.org/openpyxl/openpyxl/issue/442>`_ Styles cannot be copied
-
 
 
 2.2.0 (2015-03-11)